more explore page fixes
